@@ -63,32 +63,37 @@
 
       <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
       <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-      <maven.compiler.source>1.8</maven.compiler.source>
-      <maven.compiler.target>1.8</maven.compiler.target>
+      <maven.compiler.source>11</maven.compiler.source>
+      <maven.compiler.target>11</maven.compiler.target>
+
+      <maven-profiler-report-directory>${project.build.directory}/maven-profiler</maven-profiler-report-directory>
 
       <plugin.build-helper-maven-plugin.version>3.0.0</plugin.build-helper-maven-plugin.version>
       <plugin.docker-maven-plugin.version>0.23.0</plugin.docker-maven-plugin.version>
       <plugin.exec-maven-plugin.version>1.5.0</plugin.exec-maven-plugin.version>
       <plugin.git-commit-id-plugin.version>2.2.6</plugin.git-commit-id-plugin.version>
-      <plugin.maven-compiler-plugin.version>3.7.0</plugin.maven-compiler-plugin.version>
+      <plugin.maven-compiler-plugin.version>3.8.0</plugin.maven-compiler-plugin.version>
+      <plugin.maven-enforcer-plugin.version>3.0.0-M2</plugin.maven-enforcer-plugin.version>
       <plugin.maven-gpg-plugin.version>1.6</plugin.maven-gpg-plugin.version>
-      <plugin.maven-javadoc-plugin.version>2.10.4</plugin.maven-javadoc-plugin.version>
+      <plugin.maven-javadoc-plugin.version>3.0.1</plugin.maven-javadoc-plugin.version>
       <plugin.maven-release-plugin.version>2.5.3</plugin.maven-release-plugin.version>
-      <plugin.maven-resources-plugin.version>3.0.2</plugin.maven-resources-plugin.version>
+      <plugin.maven-resources-plugin.version>3.1.0</plugin.maven-resources-plugin.version>
       <plugin.maven-source-plugin.version>3.0.1</plugin.maven-source-plugin.version>
       <plugin.nexus-staging-maven-plugin.version>1.6.8</plugin.nexus-staging-maven-plugin.version>
-      <plugin.maven-surefire-plugin.version>2.19.1</plugin.maven-surefire-plugin.version>
+      <plugin.maven-surefire-plugin.version>3.0.0-M2</plugin.maven-surefire-plugin.version>
 
-      <dependency.assertj.version>3.8.0</dependency.assertj.version>
-      <dependency.jackson.version>2.9.2</dependency.jackson.version>
+      <dependency.assertj.version>3.11.1</dependency.assertj.version>
+      <dependency.jackson.version>2.9.8</dependency.jackson.version>
       <dependency.javax-annotation.version>1.3.2</dependency.javax-annotation.version>
-      <dependency.jedis.version>2.9.0</dependency.jedis.version>
-      <dependency.junit.version>4.12</dependency.junit.version>
-      <dependency.log4j.version>2.9.1</dependency.log4j.version>
-      <dependency.mockito.version>2.11.0</dependency.mockito.version>
+      <dependency.javax-annotation.version>1.3.2</dependency.javax-annotation.version>
+      <dependency.jedis.version>2.10.1</dependency.jedis.version>
+      <dependency.junit-platform.version>1.2.0</dependency.junit-platform.version>
+      <dependency.junit-jupiter.version>5.3.2</dependency.junit-jupiter.version>
+      <dependency.log4j.version>2.11.1</dependency.log4j.version>
+      <dependency.mockito.version>2.23.4</dependency.mockito.version>
       <dependency.slf4j.version>1.7.25</dependency.slf4j.version>
-      <dependency.spring.version>5.0.1.RELEASE</dependency.spring.version>
-      <dependency.spring-data-redis.version>2.0.1.RELEASE</dependency.spring-data-redis.version>
+      <dependency.spring.version>5.1.3.RELEASE</dependency.spring.version>
+      <dependency.spring-data-redis.version>2.1.3.RELEASE</dependency.spring-data-redis.version>
 
       <integrationTest.redisPort>16379</integrationTest.redisPort>
       <integrationTest.sentinelPort>26379</integrationTest.sentinelPort>
