@@ -1,15 +1,16 @@
 package com.s24.redjob.client;
 
-import org.springframework.beans.factory.FactoryBean;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-
 import com.s24.redjob.AbstractDao;
+import com.s24.redjob.RedJobRedisConnectionFactory;
 import com.s24.redjob.channel.ChannelDaoImpl;
 import com.s24.redjob.lock.LockDaoImpl;
 import com.s24.redjob.queue.FifoDaoImpl;
 import com.s24.redjob.worker.WorkerDaoImpl;
 import com.s24.redjob.worker.json.ExecutionRedisSerializer;
+import org.springframework.beans.factory.FactoryBean;
+import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
 
 /**
  * {@link FactoryBean} for easy creation of a {@link Client}.
@@ -83,6 +84,8 @@ public class ClientFactoryBean implements FactoryBean<Client>, InitializingBean 
    /**
     * {@link RedisConnectionFactory} to access Redis.
     */
+   @RedJobRedisConnectionFactory
+   @Autowired
    public void setConnectionFactory(RedisConnectionFactory connectionFactory) {
       workerDao.setConnectionFactory(connectionFactory);
       fifoDao.setConnectionFactory(connectionFactory);
