@@ -1,29 +1,25 @@
 package com.s24.redjob.worker.json;
 
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.s24.redjob.mockito.EnableMockito;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 
 import java.time.Instant;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnit;
-import org.mockito.junit.MockitoRule;
-
-import com.fasterxml.jackson.core.JsonGenerator;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 
 /**
  * Test for {@link InstantSerializer}.
  */
-public class InstantSerializerTest {
-   @Rule
-   public MockitoRule mockitoRule = MockitoJUnit.rule();
+@EnableMockito
+class InstantSerializerTest {
    @Mock
    private JsonGenerator jgen;
 
    @Test
-   public void serialize() throws Exception {
+   void serialize() throws Exception {
       InstantSerializer serializer = new InstantSerializer();
 
       serializer.serialize(Instant.ofEpochSecond(12345678L), jgen, null);
